chore: run linter and formatter

+++ server/src/cli/handlers/director/cli_handler_director.cpp
@@ -39,9 +39,9 @@ rtype::server::BaseContext CliHandlerDirector::BuildCtx() {
   Env env(envPath);
 
   auto name = env.Get<std::string>("NAME");
-  std::size_t port = static_cast<std::size_t>(env.Get<long>("PORT"));
-  std::size_t maxGames = static_cast<std::size_t>(env.Get<long>("MAX_GAMES"));
-  std::size_t ticks = static_cast<std::size_t>(env.Get<long>("TICKS"));
+  std::size_t port = static_cast<std::size_t>(env.Get<int>("PORT"));
+  std::size_t maxGames = static_cast<std::size_t>(env.Get<int>("MAX_GAMES"));
+  std::size_t ticks = static_cast<std::size_t>(env.Get<int>("TICKS"));
   DirectorCtxProps props = DirectorCtxProps(maxGames, ticks);
 
   return {name, port, kDirector, props};
